+++ CHANGELOG.rst
@@ -1,3 +1,8 @@
+UNRELEASED - 2017-12-21
+=======================
+
+- Add option to use different AWS settings through ``AWS_SES_ACCESS_KEY_ID``, ``AWS_SES_SECRET_ACCESS_KEY``, ``AWS_SES_REGION``
+
 1.0.0 - 2017-12-07
 ==================